requirements.txt
flask_wtf
flask_sqlalchemy
wtforms
flask_bcrypt
flask_login
flask_admin
# flask_google_login
oauthlib
Flask-FontAwesome
email-validator
soundfile
google-auth
google-auth-oauthlib
CacheControl
waitress
flask-cors
